Deprecate sendPushNotification method due to FCM upstream messaging decommission
The sendPushNotification method has been marked as deprecated since FCM upstream messaging will be decommissioned in June 2024. Also, developers are advised to implement a REST API on the server to handle device-to-server communication. For now only the server can then use the Firebase Admin SDK to send notifications to devices.